Fix: rename checkin controller, repository, entity
