+++ backend/src/main/java/z9/second/domain/checkin/dto/CheckInResponseDto.java
@@ -3,7 +3,7 @@ package z9.second.domain.checkin.dto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import z9.second.model.schedules.SchedulesCheckInEntity;
+import z9.second.model.checkIn.CheckInEntity;
 
 public class CheckInResponseDto {
     @Getter
@@ -15,27 +15,12 @@ public class CheckInResponseDto {
         private final Long userId;
         private final Boolean checkIn;
 
-        public static ResponseData from(SchedulesCheckInEntity schedulesCheckInEntity) {
+        public static ResponseData from(CheckInEntity schedulesCheckInEntity) {
             return ResponseData.builder()
                     .checkInId(schedulesCheckInEntity.getId())
                     .scheduleId(schedulesCheckInEntity.getSchedules().getId())
                     .userId(schedulesCheckInEntity.getUserId())
                     .checkIn(schedulesCheckInEntity.isCheckIn())
-                    .build();
-        }
-    }
-
-    @Getter
-    @Builder
-    @AllArgsConstructor
-    public static class ListResponseData {   // 참석자 목록 조회를 위한 응답 DTO
-        private final Long userId;
-        private final Boolean checkIn;
-
-        public static ListResponseData from(SchedulesCheckInEntity entity) {
-            return ListResponseData.builder()
-                    .userId(entity.getUserId())
-                    .checkIn(entity.isCheckIn())
                     .build();
         }
     }
