+++ backend/src/main/java/z9/second/domain/checkin/service/CheckInService.java
@@ -3,6 +3,5 @@ package z9.second.domain.checkin.service;
 import z9.second.domain.checkin.dto.CheckInRequestDto;
 
 public interface CheckInService {
-
-    void createCheckIn(Long userId, CheckInRequestDto.CreateCheckIn requestDto);
+    void CheckIn(Long userId, CheckInRequestDto.CheckInDto requestDto);
 }
