+++ backend/src/main/java/z9/second/domain/checkin/service/CheckInServiceImpl.java
@@ -7,8 +7,8 @@ import z9.second.domain.checkin.dto.CheckInRequestDto;
 import z9.second.domain.classes.repository.ClassUserRepository;
 import z9.second.global.exception.CustomException;
 import z9.second.global.response.ErrorCode;
-import z9.second.model.schedules.SchedulesCheckInEntity;
-import z9.second.model.schedules.SchedulesCheckInEntityRepository;
+import z9.second.model.checkIn.CheckInEntity;
+import z9.second.model.checkIn.CheckInEntityRepository;
 import z9.second.model.schedules.SchedulesEntity;
 import z9.second.model.schedules.SchedulesRepository;
 
@@ -16,28 +16,27 @@ import z9.second.model.schedules.SchedulesRepository;
 @RequiredArgsConstructor
 public class CheckInServiceImpl implements CheckInService {
 
-    private final SchedulesCheckInEntityRepository schedulesCheckInEntityRepository;
+    private final CheckInEntityRepository checkInEntityRepository;
     private final SchedulesRepository schedulesRepository;
     private final ClassUserRepository classUserRepository;
 
     @Transactional
     @Override
-    public void createCheckIn(Long userId, CheckInRequestDto.CreateCheckIn requestDto) {
+    public void CheckIn(Long userId, CheckInRequestDto.CheckInDto requestDto) {
         SchedulesEntity findSchedulesEntity = schedulesRepository.findById(requestDto.getScheduleId())
                 .orElseThrow(() -> new CustomException(ErrorCode.SCHEDULE_NOT_FOUND));
-
         if(!classUserRepository.existsByClassesAndUserId(findSchedulesEntity.getClasses(), userId)) {
             throw new CustomException(ErrorCode.CLASS_NOT_EXISTS_MEMBER);
         }
 
         // 모임에 가입된 회원만이, 모임 참석 유무를 결정할 수 있다.
-        SchedulesCheckInEntity newSchedulesCheckIn = SchedulesCheckInEntity
+        CheckInEntity newSchedulesCheckIn = CheckInEntity
                 .builder()
                 .schedules(findSchedulesEntity)
                 .userId(userId)
                 .checkIn(requestDto.getCheckIn())
                 .build();
 
-        schedulesCheckInEntityRepository.save(newSchedulesCheckIn);
+        checkInEntityRepository.save(newSchedulesCheckIn);
     }
 }
