+++ backend/src/main/java/z9/second/model/checkIn/CheckInEntityRepository.java
@@ -1,0 +1,7 @@
+package z9.second.model.checkIn;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CheckInEntityRepository extends JpaRepository<CheckInEntity, Long> {
+
+}
