--- backend/src/main/java/z9/second/model/schedules/SchedulesCheckInEntityRepository.java
@@ -1,7 +1,0 @@
-package z9.second.model.schedules;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SchedulesCheckInEntityRepository extends JpaRepository<SchedulesCheckInEntity, Long> {
-
-}
