+++ backend/src/main/java/z9/second/model/schedules/SchedulesEntity.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import z9.second.domain.classes.entity.ClassEntity;
 import z9.second.model.BaseEntity;
+import z9.second.model.checkIn.CheckInEntity;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -35,5 +36,5 @@ public class SchedulesEntity extends BaseEntity {
 
     @Builder.Default
     @OneToMany(mappedBy = "schedules", cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<SchedulesCheckInEntity> checkins = new ArrayList<>();
+    private List<CheckInEntity> checkins = new ArrayList<>();
 }
